fix(frontend): add generateMetadata for tournament and match dynamic pages
Tournament and match detail pages were "use client" components, which
Next.js does not allow to export generateMetadata, so every dynamic
route shared the app-wide default title/description with no per-page
SEO metadata. This mirrors the pattern already used for the profile
page: split each route into a server page.tsx that exports
generateMetadata (looking up mock tournament/match data for title,
description, and OG tags) and a sibling client component
(TournamentDetailsPageClient / MatchHubPageClient) that keeps the
original client-rendered UI and data-fetching logic unchanged. The
non-locale matches/[id] proxy route now also re-exports
generateMetadata, matching the existing profile proxy.
