+++ frontend/src/app/matches/[id]/page.tsx
@@ -1,2 +1,2 @@
 // Proxy re-export for tests
-export { default } from "@/app/[locale]/matches/[id]/page";
+export { default, generateMetadata } from "@/app/[locale]/matches/[id]/page";
